fix: handle date format change
When testing DocumentDB trigger, I found in the cdc event messages, the date format is changed to an object
type as "$date": "2025-04-29T00:00:00Z"}, so made some changes on the deserializer.
Beside, I tested by updating data in the docDB collection directly and noticed UPDATE operation type is not
handled, so added it in.

TIS21-6275