fix: XRay segment errors
Errors are thrown from `direct:` routes because the existing
`NoopTracingStrategy` does not propagate the xray context. Resulting in
an error like "Beginning new segment while another segment exists in
the segment context.".

Update the tracing strategy to `TraceAnnotatedTracingStreategy` this
ensure propagation of the context, however as there are no annotations
used in this project the additional trace data is minimised.
The alternative would be to use `TraceAllTracingStrategy`.

NO-TICKET